// frontend-elm/src/generated/session_service.ts
export interface CreateSessionRequest {
    /**
     * @generated from protobuf field: string player_name = 1
     */
    playerName: string;
    /**
     * @generated from protobuf field: int32 max_players = 2
     */
    maxPlayers: number;
    /**
     * @generated from protobuf field: string game_mode = 3
     */
    gameMode: string;
}
// ...
export interface JoinSessionRequest {
    /**
     * @generated from protobuf field: string session_code = 1
     */
    sessionCode: string;
    /**
     * @generated from protobuf field: string player_name = 2
     */
    playerName: string;
}
// ...
export interface SetReadyRequest {
    /**
     * @generated from protobuf field: string session_id = 1
     */
    sessionId: string;
    /**
     * @generated from protobuf field: string player_id = 2
     */
    playerId: string;
    /**
     * @generated from protobuf field: bool ready = 3
     */
    ready: boolean;
}
// ...
export interface GetSessionStateRequest {
    /**
     * @generated from protobuf field: string session_id = 1
     */
    sessionId: string;
}
// ...
class CreateSessionRequest$Type extends MessageType<CreateSessionRequest> {
    constructor() {
        super("takeiteasygame.v1.CreateSessionRequest", [
            { no: 1, name: "player_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "max_players", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "game_mode", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<CreateSessionRequest>): CreateSessionRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.playerName = "";
        message.maxPlayers = 0;
        message.gameMode = "";
        if (value !== undefined)
            reflectionMergePartial<CreateSessionRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CreateSessionRequest): CreateSessionRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string player_name */ 1:
                    message.playerName = reader.string();
                    break;
                case /* int32 max_players */ 2:
                    message.maxPlayers = reader.int32();
                    break;
                case /* string game_mode */ 3:
                    message.gameMode = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: CreateSessionRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string player_name = 1; */
        if (message.playerName !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.playerName);
        /* int32 max_players = 2; */
        if (message.maxPlayers !== 0)
            writer.tag(2, WireType.Varint).int32(message.maxPlayers);
        /* string game_mode = 3; */
        if (message.gameMode !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.gameMode);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message takeiteasygame.v1.CreateSessionRequest
 */
export const CreateSessionRequest = new CreateSessionRequest$Type();
// ...
class JoinSessionRequest$Type extends MessageType<JoinSessionRequest> {
    constructor() {
        super("takeiteasygame.v1.JoinSessionRequest", [
            { no: 1, name: "session_code", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "player_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<JoinSessionRequest>): JoinSessionRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sessionCode = "";
        message.playerName = "";
        if (value !== undefined)
            reflectionMergePartial<JoinSessionRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: JoinSessionRequest): JoinSessionRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string session_code */ 1:
                    message.sessionCode = reader.string();
                    break;
                case /* string player_name */ 2:
                    message.playerName = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: JoinSessionRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string session_code = 1; */
        if (message.sessionCode !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sessionCode);
        /* string player_name = 2; */
        if (message.playerName !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.playerName);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message takeiteasygame.v1.JoinSessionRequest
 */
export const JoinSessionRequest = new JoinSessionRequest$Type();
// ...
class SetReadyRequest$Type extends MessageType<SetReadyRequest> {
    constructor() {
        super("takeiteasygame.v1.SetReadyRequest", [
            { no: 1, name: "session_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "player_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "ready", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<SetReadyRequest>): SetReadyRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sessionId = "";
        message.playerId = "";
        message.ready = false;
        if (value !== undefined)
            reflectionMergePartial<SetReadyRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SetReadyRequest): SetReadyRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string session_id */ 1:
                    message.sessionId = reader.string();
                    break;
                case /* string player_id */ 2:
                    message.playerId = reader.string();
                    break;
                case /* bool ready */ 3:
                    message.ready = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SetReadyRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string session_id = 1; */
        if (message.sessionId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sessionId);
        /* string player_id = 2; */
        if (message.playerId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.playerId);
        /* bool ready = 3; */
        if (message.ready !== false)
            writer.tag(3, WireType.Varint).bool(message.ready);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message takeiteasygame.v1.SetReadyRequest
 */
export const SetReadyRequest = new SetReadyRequest$Type();
// ...
class GetSessionStateRequest$Type extends MessageType<GetSessionStateRequest> {
    constructor() {
        super("takeiteasygame.v1.GetSessionStateRequest", [
            { no: 1, name: "session_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<GetSessionStateRequest>): GetSessionStateRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sessionId = "";
        if (value !== undefined)
            reflectionMergePartial<GetSessionStateRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetSessionStateRequest): GetSessionStateRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string session_id */ 1:
                    message.sessionId = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetSessionStateRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string session_id = 1; */
        if (message.sessionId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sessionId);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message takeiteasygame.v1.GetSessionStateRequest
 */
export const GetSessionStateRequest = new GetSessionStateRequest$Type();