// frontend-elm/src/generated/game_service.ts
export interface MakeMoveRequest {
    /**
     * @generated from protobuf field: string session_id = 1
     */
    sessionId: string;
    /**
     * @generated from protobuf field: string player_id = 2
     */
    playerId: string;
    /**
     * @generated from protobuf field: string move_data = 3
     */
    moveData: string; // JSON du mouvement
    /**
     * @generated from protobuf field: int64 timestamp = 4
     */
    timestamp: bigint;
}
// ...
export interface GetGameStateRequest {
    /**
     * @generated from protobuf field: string session_id = 1
     */
    sessionId: string;
}
// ...
export interface GetAiMoveRequest {
    /**
     * @generated from protobuf field: string tile_code = 1
     */
    tileCode: string; // Code de la tuile (ex: "168")
    /**
     * @generated from protobuf field: repeated string board_state = 2
     */
    boardState: string[]; // État du plateau IA (19 cases, "" = vide)
    /**
     * @generated from protobuf field: repeated int32 available_positions = 3
     */
    availablePositions: number[]; // Positions disponibles
    /**
     * @generated from protobuf field: int32 turn_number = 4
     */
    turnNumber: number; // Numéro du tour (0-18)
}
// ...
class MakeMoveRequest$Type extends MessageType<MakeMoveRequest> {
    constructor() {
        super("takeiteasygame.v1.MakeMoveRequest", [
            { no: 1, name: "session_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "player_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "move_data", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "timestamp", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ }
        ]);
    }
    create(value?: PartialMessage<MakeMoveRequest>): MakeMoveRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sessionId = "";
        message.playerId = "";
        message.moveData = "";
        message.timestamp = 0n;
        if (value !== undefined)
            reflectionMergePartial<MakeMoveRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: MakeMoveRequest): MakeMoveRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string session_id */ 1:
                    message.sessionId = reader.string();
                    break;
                case /* string player_id */ 2:
                    message.playerId = reader.string();
                    break;
                case /* string move_data */ 3:
                    message.moveData = reader.string();
                    break;
                case /* int64 timestamp */ 4:
                    message.timestamp = reader.int64().toBigInt();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: MakeMoveRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string session_id = 1; */
        if (message.sessionId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sessionId);
        /* string player_id = 2; */
        if (message.playerId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.playerId);
        /* string move_data = 3; */
        if (message.moveData !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.moveData);
        /* int64 timestamp = 4; */
        if (message.timestamp !== 0n)
            writer.tag(4, WireType.Varint).int64(message.timestamp);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message takeiteasygame.v1.MakeMoveRequest
 */
export const MakeMoveRequest = new MakeMoveRequest$Type();
// ...
class GetGameStateRequest$Type extends MessageType<GetGameStateRequest> {
    constructor() {
        super("takeiteasygame.v1.GetGameStateRequest", [
            { no: 1, name: "session_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<GetGameStateRequest>): GetGameStateRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sessionId = "";
        if (value !== undefined)
            reflectionMergePartial<GetGameStateRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetGameStateRequest): GetGameStateRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string session_id */ 1:
                    message.sessionId = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetGameStateRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string session_id = 1; */
        if (message.sessionId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sessionId);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message takeiteasygame.v1.GetGameStateRequest
 */
export const GetGameStateRequest = new GetGameStateRequest$Type();
// ...
class GetAiMoveRequest$Type extends MessageType<GetAiMoveRequest> {
    constructor() {
        super("takeiteasygame.v1.GetAiMoveRequest", [
            { no: 1, name: "tile_code", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "board_state", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "available_positions", kind: "scalar", repeat: 1 /*RepeatType.PACKED*/, T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "turn_number", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<GetAiMoveRequest>): GetAiMoveRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.tileCode = "";
        message.boardState = [];
        message.availablePositions = [];
        message.turnNumber = 0;
        if (value !== undefined)
            reflectionMergePartial<GetAiMoveRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetAiMoveRequest): GetAiMoveRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string tile_code */ 1:
                    message.tileCode = reader.string();
                    break;
                case /* repeated string board_state */ 2:
                    message.boardState.push(reader.string());
                    break;
                case /* repeated int32 available_positions */ 3:
                    if (wireType === WireType.LengthDelimited)
                        for (let e = reader.int32() + reader.pos; reader.pos < e;)
                            message.availablePositions.push(reader.int32());
                    else
                        message.availablePositions.push(reader.int32());
                    break;
                case /* int32 turn_number */ 4:
                    message.turnNumber = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetAiMoveRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string tile_code = 1; */
        if (message.tileCode !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.tileCode);
        /* repeated string board_state = 2; */
        for (let i = 0; i < message.boardState.length; i++)
            writer.tag(2, WireType.LengthDelimited).string(message.boardState[i]);
        /* repeated int32 available_positions = 3; */
        if (message.availablePositions.length) {
            writer.tag(3, WireType.LengthDelimited).fork();
            for (let i = 0; i < message.availablePositions.length; i++)
                writer.int32(message.availablePositions[i]);
            writer.join();
        }
        /* int32 turn_number = 4; */
        if (message.turnNumber !== 0)
            writer.tag(4, WireType.Varint).int32(message.turnNumber);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message takeiteasygame.v1.GetAiMoveRequest
 */
export const GetAiMoveRequest = new GetAiMoveRequest$Type();